+++ src/Products/PluginIndexes/CompositeIndex/CompositeIndex.py
@@ -47,6 +47,31 @@ QUERY_OPERATORS = {
 MIN_COMPONENTS = 2
 
 
+def tuple_cast(kw):
+    if isinstance(kw, list):
+        kw = tuple(kw)
+    elif not isinstance(kw, tuple):
+        kw = (kw,)
+    return kw
+
+
+def collect(recs, not_cid=None):
+    kw_list = []
+    for c_id, rec in recs:
+        if c_id != not_cid:
+            keys = list(rec.keys)
+            if not_cid is not None:
+                keys.extend(rec.get('not', []))
+        elif c_id == not_cid:
+            keys = rec.get('not', [])
+        if not keys:
+            continue
+        keys = tuple_cast(keys)
+        kw = tuple([(c_id, k) for k in keys])
+        kw_list.append(kw)
+    return kw_list
+
+
 class ComponentMapping(PersistentMapping):
     """A persistent wrapper for mapping objects
     recording the order in which items are added. """
@@ -163,7 +188,7 @@ class CompositeIndex(KeywordIndex):
          'action': 'manage_browse'},
     )
 
-    query_options = ('query', 'operator')
+    query_options = ('query', 'operator', 'not')
 
     def __init__(self, id, ignore_ex=None, call_methods=None,
                  extra=None, caller=None):
@@ -322,15 +347,25 @@ class CompositeIndex(KeywordIndex):
         # collect components matching query attributes
         # and check them for completeness
         c_records = []
+
+        # component ids containing 'not' operator
+        not_cids = []
+
         for c in components:
             query_options = QUERY_OPTIONS[c.meta_type]
             query_operators = QUERY_OPERATORS[c.meta_type]
             rec = IndexQuery(query, c.id, query_options,
                              query_operators[0], query_operators[1])
+            opr = None
 
-            # not supported: 'not' parameter
-            not_parm = rec.get('not', None)
-            if not rec.keys and not_parm:
+            # not supported: 'range' parameter
+            range_parm = rec.get('range', None)
+            if range_parm:
+                opr = 'range'
+            if rec.get('usage', None):
+                # see if any usage params are sent to field
+                opr = rec.usage.lower().split(':')
+            if opr == 'range':
                 continue
 
             # not supported: 'and' operator
@@ -345,29 +380,43 @@ class CompositeIndex(KeywordIndex):
             if c.meta_type == 'BooleanIndex':
                 rec.keys = [int(bool(v)) for v in rec.keys[:]]
 
+            # rec with 'not' parameter
+            not_parm = rec.get('not', None)
+            if not_parm:
+                # not supported: 'pure not'
+                if len(rec.keys) == 0:
+                    continue
+                not_cids.append(c.id)
+                if c.meta_type == 'BooleanIndex':
+                    not_parm = [int(bool(v)) for v in not_parm[:]]
+                    rec.set('not', not_parm)
+
             c_records.append((c.id, rec))
 
         # return if less than MIN_COMPONENTS query attributes were caught
         if len(c_records) < MIN_COMPONENTS:
             return query
 
-        kw_list = []
-        for c_id, rec in c_records:
-            kw = rec.keys
-            if not kw:
-                continue
-            if isinstance(kw, list):
-                kw = tuple(kw)
-            elif not isinstance(kw, tuple):
-                kw = (kw,)
-            kw = tuple([(c_id, k) for k in kw])
-            kw_list.append(kw)
-
+        records = ()
+        kw_list = collect(c_records)
         # permute keyword list
         records = tuple(product(*kw_list))
 
+        not_records = set()
+        for c_id in not_cids:
+            kw_list = collect(c_records, not_cid=c_id)
+            if kw_list:
+                not_records.update(product(*kw_list))
+        # permute keyword list for 'not' operator
+        not_records = tuple(not_records)
+
         # substitute matching query attributes as composite index
-        cquery.update({self.id: {'query': records}})
+        if records and not_records:
+            cquery.update({self.id: {'query': records, 'not': not_records}})
+        elif records:
+            cquery.update({self.id: {'query': records}})
+        elif not_records:
+            cquery.update({self.id: {'not': not_records}})
 
         # delete original matching query attributes from query
         for c_id, rec in c_records: