+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 4.5 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Fix rewriting of query to avoid wrong optimization of CompositeIndex.
+  (`#59 <https://github.com/zopefoundation/Products.ZCatalog/issues/59>`_)
 
 
 4.4 (2019-03-08)